backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0
python-dotenv>=1.0.0
google-genai>=0.8.0
supabase>=2.4.0
pdfplumber>=0.11.0
rank-bm25>=0.2.2
python-multipart>=0.0.9
httpx>=0.27.0
